Fix importing `@tauri-apps/api` in Node.js
